Refactored states, added state set name
